+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - [Check for correct update operation access in location hierarchy form #800](https://github.com/farmOS/farmOS/pull/800)
 - [Update to itamair/geophp 1.6 #801](https://github.com/farmOS/farmOS/pull/801) to fix deprecated function warnings.
+- [Fix saving inventory quick form configuration with existing units term #803](https://github.com/farmOS/farmOS/pull/803)
 
 ## [3.1.1] 2024-02-07